+++ rollup.config.js
@@ -6,25 +6,49 @@ import packageJSON from './package.json';
 
 const externals = Object.keys(packageJSON.dependencies ? packageJSON.dependencies : {});
 
-export default {
-  input: "src/index.ts",
-  treeshake: true,
-  external: externals,
-  output: [
-    {
-      file: "dist/index.js",
-      format: "es"
-    }
-  ],
-  plugins: [
-    json(),
-    nodeResolve({
-      preferBuiltins: true
-    }),
-    commonjs({
-      include: "node_modules/**",
-      sourceMap: true
-    }),
-    typescript()
-  ]
-}
+export default [
+  {
+    input: "src/index.ts",
+    treeshake: true,
+    external: externals,
+    output: [
+      {
+        file: "dist/utility.es.js",
+        format: "es"
+      }
+    ],
+    plugins: [
+      json(),
+      nodeResolve({
+        preferBuiltins: true
+      }),
+      commonjs({
+        include: "node_modules/**",
+        sourceMap: true
+      }),
+      typescript()
+    ]
+  },
+  {
+    input: "src/index.ts",
+    treeshake: true,
+    external: externals,
+    output: [
+      {
+        file: "dist/utility.cjs.js",
+        format: "cjs"
+      }
+    ],
+    plugins: [
+      json(),
+      nodeResolve({
+        preferBuiltins: true
+      }),
+      commonjs({
+        include: "node_modules/**",
+        sourceMap: true
+      }),
+      typescript()
+    ]
+  }
+]