+++ src/interpolation/index.ts
@@ -1,3 +1,9 @@
+type InterpolationOptions = {
+  currentTime: number,
+  startAt: number,
+  finishAt: number,
+  durationTime: number
+}
 /**
  *
  * @class Interpolation
@@ -34,10 +40,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeInCubic
    * @description Interpolation function of cubic easing with accelerating from zero velocity
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-   easeInCubic (options: Kenift.Utility.InterpolationOptions): number {
+   easeInCubic (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime;
 
     return options.finishAt*options.currentTime*options.currentTime*options.currentTime + options.startAt;
@@ -71,10 +77,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeInOutCubic
    * @description Interpolation function of cubic easing with acceleration until halfway, then deceleration
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-   easeInOutCubic (options: Kenift.Utility.InterpolationOptions): number {
+   easeInOutCubic (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime/2;
     if (options.currentTime < 1) return options.finishAt/2*options.currentTime*options.currentTime*options.currentTime + options.startAt;
     options.currentTime -= 2;
@@ -109,10 +115,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeOutCubic
    * @description Interpolation function of cubic easing with decelerating to zero velocity
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-   easeOutCubic (options: Kenift.Utility.InterpolationOptions): number {
+   easeOutCubic (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime;
     options.currentTime--;
     return options.finishAt*(options.currentTime*options.currentTime*options.currentTime + 1) + options.startAt;
@@ -146,9 +152,9 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function linear
    * @description Interpolation function with no easing and no acceleration effect
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    */
-  linear (options: Kenift.Utility.InterpolationOptions): number {
+  linear (options: InterpolationOptions): number {
     return options.finishAt*options.currentTime/options.durationTime + options.startAt;
   }
   /**
@@ -180,10 +186,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easyInOutQuad
    * @description Interpolation function of quad easing with acceleration until halfway, then deceleration
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeInOutQuad (options: Kenift.Utility.InterpolationOptions): number {
+  easeInOutQuad (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime/2;
     if (options.currentTime < 1) return options.finishAt/2*options.currentTime*options.currentTime + options.startAt;
     options.currentTime--;
@@ -218,10 +224,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeInQuad
    * @description Interpolation function of quad easing with accelerating from zero velocity
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  public easeInQuad (options: Kenift.Utility.InterpolationOptions): number {
+  public easeInQuad (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime;
     return options.finishAt*options.currentTime*options.currentTime + options.startAt;
   }
@@ -254,10 +260,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeOutQuad
    * @description Interpolation function of quad easing with decelerating to zero velocity
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeOutQuad (options: Kenift.Utility.InterpolationOptions): number {
+  easeOutQuad (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime;
     return -options.finishAt * options.currentTime*(options.currentTime-2) + options.startAt;
   }
@@ -290,10 +296,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeInQuart
    * @description Interpolation function of quart easing with accelerating from zero velocity
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeInQuart (options: Kenift.Utility.InterpolationOptions): number {
+  easeInQuart (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime;
     return options.finishAt*options.currentTime*options.currentTime*options.currentTime*options.currentTime + options.startAt;
   }
@@ -326,10 +332,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeOutQuart
    * @description Interpolation function of quart easing with decelerating to zero velocity
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeOutQuart (options: Kenift.Utility.InterpolationOptions): number {
+  easeOutQuart (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime;
     options.currentTime--;
     return -options.finishAt * (options.currentTime*options.currentTime*options.currentTime*options.currentTime - 1) + options.startAt;
@@ -363,10 +369,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeInOutQuart
    * @description Interpolation function of quart easing with acceleration until halfway, then deceleration
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeInOutQuart (options: Kenift.Utility.InterpolationOptions): number {
+  easeInOutQuart (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime/2;
     if (options.currentTime < 1) return options.finishAt/2*options.currentTime*options.currentTime*options.currentTime*options.currentTime + options.startAt;
     options.currentTime -= 2;
@@ -401,10 +407,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeInQuint
    * @description Interpolation function of quint easing with accelerating from zero velocity
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeInQuint (options: Kenift.Utility.InterpolationOptions): number {
+  easeInQuint (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime;
     return options.finishAt*options.currentTime*options.currentTime*options.currentTime*options.currentTime*options.currentTime + options.startAt;
   }
@@ -437,10 +443,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeOutQuint
    * @description Interpolation function of quint easing with decelerating to zero velocity
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeOutQuint (options: Kenift.Utility.InterpolationOptions): number {
+  easeOutQuint (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime;
     options.currentTime--;
     return options.finishAt*(options.currentTime*options.currentTime*options.currentTime*options.currentTime*options.currentTime + 1) + options.startAt;
@@ -474,10 +480,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeInOutQuint
    * @description Interpolation function of quint easing with acceleration until halfway, then deceleration
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeInOutQuint (options: Kenift.Utility.InterpolationOptions): number {
+  easeInOutQuint (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime/2;
     if (options.currentTime < 1) return options.finishAt/2*options.currentTime*options.currentTime*options.currentTime*options.currentTime*options.currentTime + options.startAt;
     options.currentTime -= 2;
@@ -512,10 +518,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeInSine
    * @description Interpolation function of sinusoidal easing with accelerating from zero velocity
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeInSine (options: Kenift.Utility.InterpolationOptions): number {
+  easeInSine (options: InterpolationOptions): number {
     return -options.finishAt * Math.cos(options.currentTime/options.durationTime * (Math.PI/2)) + options.finishAt + options.startAt;
   }
   /**
@@ -547,10 +553,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeOutSine
    * @description Interpolation function of sinusoidal easing with decelerating to zero velocity
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeOutSine (options: Kenift.Utility.InterpolationOptions): number {
+  easeOutSine (options: InterpolationOptions): number {
     return options.finishAt * Math.sin(options.currentTime/options.durationTime * (Math.PI/2)) + options.startAt;
   }
   /**
@@ -582,10 +588,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeInOutSine
    * @description Interpolation function of sinusoidal easing with accelerating until halfway, then decelerating
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeInOutSine (options: Kenift.Utility.InterpolationOptions): number {
+  easeInOutSine (options: InterpolationOptions): number {
     return -options.finishAt/2 * (Math.cos(Math.PI*options.currentTime/options.durationTime) - 1) + options.startAt;
   }
   /**
@@ -617,10 +623,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeInExpo
    * @description Interpolation function of exponential easing with accelerating from zero velocity
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeInExpo (options: Kenift.Utility.InterpolationOptions): number {
+  easeInExpo (options: InterpolationOptions): number {
     return options.finishAt * Math.pow( 2, 10 * (options.currentTime/options.durationTime - 1) ) + options.startAt;
   }
   /**
@@ -652,10 +658,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeOutExpo
    * @description Interpolation function of exponential easing with decelerating to zero velocity
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeOutExpo (options: Kenift.Utility.InterpolationOptions): number {
+  easeOutExpo (options: InterpolationOptions): number {
     return options.finishAt * ( -Math.pow( 2, -10 * options.currentTime/options.durationTime ) + 1 ) + options.startAt;
   }
   /**
@@ -687,10 +693,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeInOutExpo
    * @description Interpolation function of exponential easing with accelerating until halfway, then decelerating
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeInOutExpo (options: Kenift.Utility.InterpolationOptions): number {
+  easeInOutExpo (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime/2;
     if (options.currentTime < 1) return options.finishAt/2 * Math.pow( 2, 10 * (options.currentTime - 1) ) + options.startAt;
     options.currentTime--;
@@ -725,10 +731,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeInCirc
    * @description Interpolation function of circular easing with accelerating from zero velocity
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeInCirc (options: Kenift.Utility.InterpolationOptions): number {
+  easeInCirc (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime;
     return -options.finishAt * (Math.sqrt(1 - options.currentTime*options.currentTime) - 1) + options.startAt;
   }
@@ -761,10 +767,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeOutCirc
    * @description Interpolation function of circular easing with decelerating to zero velocity
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeOutCirc (options: Kenift.Utility.InterpolationOptions): number {
+  easeOutCirc (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime;
     options.currentTime--;
     return options.finishAt * Math.sqrt(1 - options.currentTime*options.currentTime) + options.startAt;
@@ -798,10 +804,10 @@ export const Interpolation = new class Interpolation {
    * @public
    * @function easeInOutCirc
    * @description Interpolation function of circular easing with acceleration until halfway, then deceleration
-   * @param {Kenift.Utility.InterpolationOptions} options
+   * @param {InterpolationOptions} options
    * @returns {number}
    */
-  easeInOutCirc (options: Kenift.Utility.InterpolationOptions): number {
+  easeInOutCirc (options: InterpolationOptions): number {
     options.currentTime /= options.durationTime/2;
     if (options.currentTime < 1) return -options.finishAt/2 * (Math.sqrt(1 - options.currentTime*options.currentTime) - 1) + options.startAt;
     options.currentTime -= 2;